django_api/requirements.txt
Django
mysqlclient
pymysql
djangorestframework
woocommerce
django-cors-headers
cryptography